@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { Howl, Howler } from 'howler';
+import { Editor } from './Editor/Editor';
 
 import logo from './logo.svg';
 import './App.css';
@@ -50,16 +51,15 @@ class App extends Component {
     render() {
     return (
       <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-            <p className="App-intro">
-                <button onClick={this.playMusic}>PLAY</button>
-                <button onClick={this.pauseMusic}>PAUSE</button>
+        <header className="App-header">Elastic Tunes</header>
+        <p className="App-intro">
+            <button onClick={this.playMusic}>PLAY</button>
+            <button onClick={this.pauseMusic}>PAUSE</button>
 
-          To get started, edit <code>src/App.js</code> and save to reload.
+        To get started, edit <code>src/App.js</code> and save to reload.
         </p>
+
+        <Editor />
       </div>
     );
   }
